+++ src/is/index.js
@@ -1,4 +1,4 @@
-import isObject, {isEmptyObject, isObjectLike} from './isObject'
+import isObject, { isEmptyObject, isObjectLike } from './isObject'
 import isArray from './isArray'
 import isEmpty from './isEmpty'
 import isEven from './isEven'
@@ -10,7 +10,7 @@ import isArrayLike from './isArrayLike'
 import isBasicType from './isBasicType'
 import isBuffer from './isBuffer'
 import isArguments from './isArguments'
-import isNil, {isNils, isNull, isNulls} from './isNil'
+import isNil, { isNils, isNull, isNulls } from './isNil'
 
 /**
  * This function evaluates if all the parameters are strings